move line storage
